requirements.txt
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
cvlib==0.2.6
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.34.1
idna==3.2
imageio==2.9.0
imutils==0.5.4
joblib==1.0.1
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
Pillow==8.3.1
progressbar==2.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
